# LATCH_GO_PORT_API_SPEC.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Latch API",
    "version": "1.0.0",
    "description": "Machine-readable spec for porting Latch Next.js /api/* routes to Go. Extension-verified. See LATCH_GO_PORT_GUIDE.md."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Go dev server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Next.js reference server"
    }
  ],
...
  "paths": {
...
    "/api/smart-account/setup-swap-rules": {
      "post": {
        "summary": "POST /api/smart-account/setup-swap-rules",
        "operationId": "_api_smart-account_setup-swap-rules_post",
        "tags": [
          "smart-account"
        ],
        "x-latch-source-file": "app/api/smart-account/setup-swap-rules/route.ts",
        "x-latch-auth": "none",
        "x-latch-env-required": [
          "BUNDLER_SECRET",
          "NEXT_PUBLIC_FACTORY_ADDRESS",
          "NEXT_PUBLIC_VERIFIER_ADDRESS",
          "NEXT_PUBLIC_WEBAUTHN_VERIFIER_ADDRESS",
          "NEXT_PUBLIC_COUNTER_ADDRESS",
          "NEXT_PUBLIC_RPC_URL",
          "NEXT_PUBLIC_NETWORK_PASSPHRASE",
          "MAINNET_RPC_URL",
          "MAINNET_NETWORK_PASSPHRASE"
        ],
        "x-latch-db-models": [],
        "x-latch-legacy-error": false,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BuildAuthTransactionResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "routerContractId": {
                          "type": "string"
                        },
                        "providerId": {
                          "type": "string"
                        },
                        "contextRuleId": {
                          "type": "integer"
                        },
                        "remainingSetupCount": {
                          "type": "integer"
                        },
                        "instructions": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BuildAuthTransactionResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "routerContractId": {
                          "type": "string"
                        },
                        "providerId": {
                          "type": "string"
                        },
                        "contextRuleId": {
                          "type": "integer"
                        },
                        "remainingSetupCount": {
                          "type": "integer"
                        },
                        "instructions": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "network",
                  "smartAccountAddress"
                ],
                "properties": {
                  "network": {
                    "$ref": "#/components/schemas/Network"
                  },
                  "smartAccountAddress": {
                    "type": "string"
                  },
                  "signerType": {
                    "$ref": "#/components/schemas/SignerType"
                  },
                  "providerId": {
                    "type": "string"
                  },
                  "routerContractId": {
                    "type": "string"
                  },
                  "publicKeyHex": {
                    "type": "string"
                  },
                  "gAddress": {
                    "type": "string"
                  },
                  "keyDataHex": {
                    "type": "string"
                  },
                  "credentialId": {
                    "type": "string"
                  },
                  "credential_id": {
                    "type": "string",
                    "description": "alias for credentialId"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/multisig/proposals/{id}/approve/webauthn": {
      "post": {
        "summary": "POST /api/multisig/proposals/{id}/approve/webauthn",
        "operationId": "_api_multisig_proposals_id_approve_webauthn_post",
        "tags": [
          "multisig"
        ],
        "x-latch-source-file": "app/api/multisig/proposals/[id]/approve/webauthn/route.ts",
        "x-latch-auth": "session+ownership",
        "x-latch-env-required": [
          "DATABASE_URL",
          "DIRECT_URL"
        ],
        "x-latch-db-models": [
          "MultisigApproval"
        ],
        "x-latch-legacy-error": true,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid",
        "description": "30-day sliding session. Auto-created on first request."
      }
    },
    "schemas": {
      "ApiErrorBody": {
        "type": "object",
        "required": [
          "error",
          "code",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "LegacyErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "SignerType": {
        "type": "string",
        "enum": [
          "passkey",
          "phantom",
          "freighter"
        ]
      },
      "Network": {
        "type": "string",
        "enum": [
          "testnet",
          "mainnet"
        ]
      },
      "ContextRuleDiscovery": {
        "type": "string",
        "enum": [
          "matched",
          "default",
          "fallback"
        ]
      },
      "SubmitMethod": {
        "type": "string",
        "enum": [
          "bundler-delegated",
          "delegated",
          "webauthn"
        ]
      },
      "BuildAuthTransactionResult": {
        "type": "object",
        "required": [
          "txXdr",
          "authEntryXdr",
          "authEntriesXdr",
          "smartAccountAuthEntryIndex",
          "delegatedNativeAuthEntryIndices",
          "delegatedNativeSignBlobPayloadsBase64",
          "delegatedGAuthEntrySynthesized",
          "contextRuleId",
          "contextRuleIds",
          "contextRuleDiscovery",
          "authDigestHex",
          "signaturePayloadHex",
          "validUntilLedger",
          "simulationResultXdr"
        ],
        "properties": {
          "txXdr": {
            "type": "string"
          },
          "authEntryXdr": {
            "type": "string"
          },
          "authEntriesXdr": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "smartAccountAuthEntryIndex": {
            "type": "integer"
          },
          "delegatedNativeAuthEntryIndices": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "delegatedNativeSignBlobPayloadsBase64": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "delegatedGAuthEntrySynthesized": {
            "type": "boolean"
          },
          "contextRuleId": {
            "type": "integer"
          },
          "contextRuleIds": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "contextRuleDiscovery": {
            "$ref": "#/components/schemas/ContextRuleDiscovery"
          },
          "authDigestHex": {
            "type": "string"
          },
          "signaturePayloadHex": {
            "type": "string"
          },
          "validUntilLedger": {
            "type": "integer"
          },
          "simulationResultXdr": {
            "type": "string"
          },
          "smartAccountAuthEntryXdr": {
            "type": "string"
          },
          "gAddressPreimageXdr": {
            "type": "string"
          },
          "gAddressEntryTemplateXdr": {
            "type": "string"
          },
          "submitMethod": {
            "$ref": "#/components/schemas/SubmitMethod"
          },
          "delegatedAuthG": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}